requirements.txt
h5py
numpy
pillow
anytree
tqdm